I have complected some of the waiter backend. The waiter now checks, every time the page loads, if the status for the order has changed then it will be printed.
As of yet I am not sure how to return this data, so for now it is printed on the local server.

apps.py:
I have renamed the class, as requested.

settings.py:
changed the name to be the new class, see apps.py. and added waiterconfig

urls.py:
added basic index for testing waiter class